% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_mermaid_project_endpoints.R
\name{get_all_mermaid_project_endpoints}
\alias{get_all_mermaid_project_endpoints}
\title{Get all endpoints from a specified MERMAID project}
\usage{
get_all_mermaid_project_endpoints(project = get_default_project(),
  limit = 50, url = base_url, token = mermaid_token())
}
\arguments{
\item{project}{A way to identify the project. Can be a project ID (passed as a character vector directly) or a single project resulting from \code{\link{get_mermaid_endpoint}} or \code{\link{search_projects}}. Defaults to the project listed via \code{get_default_project}, if available.}

\item{limit}{Number of records to get. Defaults to 50.}

\item{url}{API URL. Defaults to https://dev-api.datamermaid.org}

\item{token}{API token. Not required for unauthenticated endpoints. Get via \code{\link{mermaid_auth}}}
}
\value{
A list of tibbles, one for each of the MERMAID project endpoints (beltfishtransectmethods, beltfishes, benthiclittransectmethods, benthicpittransectmethods, benthicpits, benthictransects, collectrecords, fishbelttransects, habitatcomplexities, obsbenthiclits, obsbenthicpits, obshabitatcomplexities, obstransectbeltfishs, managements, observers, profiles, project_profiles, sampleevents, sites).
}
\description{
Get all endpoints from a specified MERMAID project
}
\examples{
test_project <- search_projects("Sharla test")
get_all_mermaid_project_endpoints(test_project)
}
