% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_get_endpoint.R
\name{get_endpoint}
\alias{get_endpoint}
\title{Get MERMAID endpoint}
\usage{
get_endpoint(
  endpoint = c("benthicattributes", "choices", "fishfamilies", "fishgenera",
    "fishspecies", "fishsizes", "managements", "projects", "projecttags", "sites"),
  limit = NULL,
  url = base_url,
  ...
)
}
\arguments{
\item{endpoint}{Endpoint}

\item{limit}{Number of records to get. Use NULL (the default) to get all records.}

\item{url}{API URL. Defaults to https://api.datamermaid.org}

\item{...}{Additional parameters used as needed}
}
\description{
Get MERMAID endpoint
}
