% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mermaid_endpoint.R
\name{get_mermaid_endpoint}
\alias{get_mermaid_endpoint}
\title{Get MERMAID Endpoint}
\usage{
get_mermaid_endpoint(endpoint = c("benthicattributes", "fishattributes",
  "fishfamilies", "fishgenera", "fishspecies", "managements", "projects",
  "sites"), limit = 50, url = base_url)
}
\arguments{
\item{endpoint}{Endpoint}

\item{limit}{Number of records to get. Defaults to 50.}

\item{url}{API URL. Defaults to https://dev-api.datamermaid.org}
}
\description{
Get MERMAID Endpoint
}
