% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_get.R
\name{get_mermaid_organization_profiles}
\alias{get_mermaid_organization_profiles}
\title{Get MERMAID organization profiles endpoint.}
\usage{
get_mermaid_organization_profiles(limit = 50, results_only = TRUE)
}
\arguments{
\item{limit}{Number of records to get. Defaults to 50.}

\item{results_only}{Whether to return the results only (the default) or to return the entire API call (contents, results, path, and response).}
}
\description{
Get MERMAID organization profiles endpoint.
}
\seealso{
Other MERMAID endpoints: \code{\link{get_mermaid_benthic_attributes}},
  \code{\link{get_mermaid_choices}},
  \code{\link{get_mermaid_fish_attributes}},
  \code{\link{get_mermaid_fish_families}},
  \code{\link{get_mermaid_fish_genera}},
  \code{\link{get_mermaid_fish_species}},
  \code{\link{get_mermaid_managements}},
  \code{\link{get_mermaid_organizations}},
  \code{\link{get_mermaid_profiles}},
  \code{\link{get_mermaid_projects}},
  \code{\link{get_mermaid_sites}}
}
\concept{MERMAID endpoints}
