% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mermaid_token.R
\name{get_mermaid_token}
\alias{get_mermaid_token}
\title{Get MERMAID API token}
\usage{
get_mermaid_token(client_id = mermaid_api_client("id"),
  client_secret = mermaid_api_client("secret"))
}
\arguments{
\item{client_id}{API client id}

\item{client_secret}{API client secret}
}
\description{
Get MERMAID API token
}
