% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_get.R
\name{get_project_managements}
\alias{get_project_managements}
\title{Get managements for a given project.}
\usage{
get_project_managements(project_id, limit = 50, results_only = TRUE)
}
\arguments{
\item{project_id}{Project ID}

\item{limit}{Number of records to get. Defaults to 50.}

\item{results_only}{Whether to return the results only (the default) or to return the entire API call (contents, results, path, and response).}
}
\description{
Get managements for a given project.
}
\seealso{
Other project endpoints: \code{\link{get_project_belt_fish_transect_methods}},
  \code{\link{get_project_beltfishes}},
  \code{\link{get_project_benthic_lit_transect_methods}},
  \code{\link{get_project_benthic_pit_transect_methods}},
  \code{\link{get_project_benthic_pits}},
  \code{\link{get_project_benthic_transects}},
  \code{\link{get_project_collect_records}},
  \code{\link{get_project_fishbelt_transects}},
  \code{\link{get_project_habitat_complexities}},
  \code{\link{get_project_obs_benthic_lits}},
  \code{\link{get_project_obs_benthic_pits}},
  \code{\link{get_project_obs_habitat_complexities}},
  \code{\link{get_project_obs_transect_beltfishs}},
  \code{\link{get_project_observers}},
  \code{\link{get_project_profiles}},
  \code{\link{get_project_project_profiles}},
  \code{\link{get_project_sample_events}},
  \code{\link{get_project_sites}}
}
\concept{project endpoints}
