% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mermaid-class.R
\name{mermaid2.0_token}
\alias{mermaid2.0_token}
\title{Generate a mermaid token}
\usage{
mermaid2.0_token(endpoint, app, scope = NULL, user_params = NULL,
  type = NULL, use_oob = getOption("httr_oob_default"),
  oob_value = NULL, as_header = TRUE, use_basic_auth = FALSE,
  cache = getOption("httr_oauth_cache"), config_init = list(),
  client_credentials = FALSE, credentials = NULL,
  query_authorize_extra = list())
}
\description{
Constructor function for objects of class [Mermaid2.0].
}
