% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_api_client.R
\name{mermaid_api_client}
\alias{mermaid_api_client}
\title{Get or set MERMAID_API_CLIENT_ID and MERMAID_API_CLIENT_SECRET}
\usage{
mermaid_api_client(value = c("id", "secret"))
}
\arguments{
\item{force}{force setting a new MERMAID API Client ID and Secret for the current environment?}
}
\value{
atomic character vector containing the MERMAID API Client ID and Secret
}
\description{
The API wrapper functions in this package all rely on a MERMAID API Client ID and Secret residing in the environment variables \code{MERMAID_API_CLIENT_ID} and \code{MERMAID_API_CLIENT_SECRET}. The
easiest way to accomplish this is to set them in the `\code{.Renviron}` file in your
home directory.
}
