% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_auth.R
\name{mermaid_auth}
\alias{mermaid_auth}
\title{Authorize \code{mermaidr}}
\usage{
mermaid_auth(token = NULL, new_user = FALSE, key = mermaid_key,
  cache = getOption("mermaidr.httr_oauth_cache"), verbose = TRUE)
}
\arguments{
\item{token}{optional; an actual token object or the path to a valid token
stored as an \code{.rds} file}

\item{new_user}{logical, defaults to \code{FALSE}. Set to \code{TRUE} if you
want to wipe the slate clean and re-authenticate with the same or different account. This disables the \code{.httr-oauth} file in current
working directory.}

\item{key}{the "Client ID" for the application}

\item{cache}{logical indicating if \code{mermaidr} should cache
credentials in the default cache file \code{.httr-oauth}}

\item{verbose}{logical; do you want informative messages?}
}
\description{
Authorize \code{mermaidr} to access the MERMAID API. You will be
directed to a web browser and asked to sign in to MERMAID Collect. By default, these user credentials are cached in a
file named \code{.httr-oauth} in the current working directory, from where
they can be automatically refreshed, as necessary.
}
\details{
Most users, most of the time, do not need to call this function
explicitly -- it will be triggered by the first action that
requires authorization. Even when called, the default arguments will often
suffice. However, when necessary, this function allows the user to

\itemize{
  \item force the creation of a new token
  \item retrieve current token as an object, for possible storage to an
  \code{.rds} file
  \item read the token from an object or from an \code{.rds} file
  \item prevent caching of credentials in \code{.httr-oauth}
}

In a direct call to \code{mermaid_auth}, the user can dictate whether interactively-obtained
credentials will be cached in \code{.httr_oauth}. If unspecified, these
arguments are controlled via options, which, if undefined at the time
\code{mermaidr} is loaded, are defined like so:

\describe{
  \item{cache}{Set to option \code{mermaidr.httr_oauth_cache}, which
  defaults to \code{TRUE}}
}

To override these defaults in persistent way, predefine one or more of
them with lines like this in a \code{.Rprofile} file:
\preformatted{
options(mermaidr.httr_oauth_cache = FALSE)
}
See \code{\link[base]{Startup}} for possible locations for this file and the
implications thereof.
}
\examples{
\dontrun{
## load/refresh existing credentials, if available
## otherwise, go to browser for authentication and authorization
mermaid_auth()

## force a new token to be obtained
mermaid_auth(new_user = TRUE)

## store token in an object and then to file
ttt <- mermaid_auth()
saveRDS(ttt, "ttt.rds")

## load a pre-existing token
mermaid_auth(token = ttt) # from an object
mermaid_auth(token = "ttt.rds") # from .rds file
}
}
\seealso{
Other auth functions: \code{\link{mermaid_deauth}}
}
\concept{auth functions}
