% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_get.R
\name{mermaid_benthic_attributes}
\alias{mermaid_benthic_attributes}
\alias{mermaid_choices}
\alias{mermaid_fish_attributes}
\alias{mermaid_fish_genera}
\alias{mermaid_fish_species}
\alias{mermaid_managements}
\alias{mermaid_organization_profiles}
\alias{mermaid_organizations}
\alias{mermaid_profiles}
\alias{mermaid_projects}
\alias{mermaid_sites}
\title{Get MERMAID endpoints.}
\usage{
mermaid_benthic_attributes(limit = 50, results_only = TRUE)

mermaid_choices(limit = 50, results_only = TRUE)

mermaid_fish_attributes(limit = 50, results_only = TRUE)

mermaid_fish_genera(limit = 50, results_only = TRUE)

mermaid_fish_species(limit = 50, results_only = TRUE)

mermaid_managements(limit = 50, results_only = TRUE)

mermaid_organization_profiles(limit = 50, results_only = TRUE)

mermaid_organizations(limit = 50, results_only = TRUE)

mermaid_profiles(limit = 50, results_only = TRUE)

mermaid_projects(limit = 50, results_only = TRUE)

mermaid_sites(limit = 50, results_only = TRUE)
}
\arguments{
\item{limit}{Number of records to get. Defaults to 50.}

\item{results_only}{Whether to return the results only (the default) or to return the entire API call (contents, results, path, and response).}
}
\description{
These wrapper functions make it easy to get overall data.
There are 12 functions:
\itemize{
  \item \code{mermaid_benthic_attributes()}
  \item \code{mermaid_choices()}
  \item \code{mermaid_fish_attributes()}
  \item \code{mermaid_fish_families()}
  \item \code{mermaid_fish_genera()}
  \item \code{mermaid_fish_species()}
  \item \code{mermaid_managements()}
  \item \code{mermaid_organization_profiles()}
  \item \code{mermaid_organizations()}
  \item \code{mermaid_profiles()}
  \item \code{mermaid_projects()}
  \item \code{mermaid_sites()}

}
}
\examples{
print("todo")
}
