% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_clean_columns.R
\name{mermaid_clean_columns}
\alias{mermaid_clean_columns}
\title{Expand columns in MERMAID output}
\usage{
mermaid_clean_columns(.data, append_column_prefix = FALSE, clean_names = TRUE)
}
\arguments{
\item{.data}{Input data}

\item{append_column_prefix}{Whether to append the df-cols name as a prefix. Defaults to FALSE}

\item{clean_names}{Whether clean the new column names into \code{snake_case} format, in case they are not clean (e.g. contain spaces, dashes, etc). Note that the name cleaning applies to *all* columns, not just ones resulting from the df-cols.}
}
\description{
Expands data-frame columns (df-cols) that come from \code{mermaidr}. Optionally cleans the column names.
}
\examples{
\dontrun{
library(dplyr)

sampleunits <- mermaid_get_my_projects() \%>\%
  head(1) \%>\%
  mermaid_get_project_data("fishbelt", "sampleunits", limit = 1) \%>\%
  select(project, biomass_kgha_by_trophic_group)

names(sampleunits)
# [1] "project" "biomass_kgha_by_trophic_group"

sampleunits <- sampleunits \%>\%
  mermaid_clean_columns()

names(sampleunits)
# [1] "project"               "piscivore"             "planktivore"
# [4] "invertivore_mobile"    "herbivore_detritivore"
}
}
