% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_countries.R
\name{mermaid_countries}
\alias{mermaid_countries}
\title{MERMAID countries}
\usage{
mermaid_countries()
}
\description{
Returns the names of countries in MERMAID to help with searching.
}
\examples{
\dontrun{
mermaid_countries()
}
}
