% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_default_project.R
\name{mermaid_set_default_project}
\alias{mermaid_set_default_project}
\alias{mermaid_get_default_project}
\title{Get or set default \code{mermaidr} MERMAID projects}
\usage{
mermaid_set_default_project(project)

mermaid_get_default_project()
}
\arguments{
\item{project}{A way to identify MERMAID project(s). Can be project IDs (passed as a character vector directly) or projects resulting from \code{\link{mermaid_get_my_projects}} or \code{\link{mermaid_search_my_projects}}.}
}
\description{
Configure default MERMAID projects. \code{mermaid_set_default_project} sets project(s) as the default to be used automatically in \code{mermaid_get_project_*} functions. \code{mermaid_get_default_project} lists the default projects.
}
\examples{
\dontrun{
test_project <- mermaid_search_my_projects("Sharla test", include_test_projects = TRUE)
mermaid_set_default_project(test_project)
mermaid_get_default_project()

# Since a default project is set, you can use e.g. mermaid_get_project_sites()
# without explicitly supplying a project
mermaid_get_project_sites()
}
}
