% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_GET.R
\name{mermaid_GET}
\alias{mermaid_GET}
\title{Get MERMAID endpoint}
\usage{
mermaid_GET(endpoint, limit = 50, url = base_url, token = NULL)
}
\arguments{
\item{endpoint}{Endpoint}

\item{limit}{Number of records to get. Defaults to 50.}

\item{url}{API URL. Defaults to https://dev-api.datamermaid.org}

\item{token}{API token. Not required for unauthenticated endpoints. Get via \code{\link{mermaid_auth}}}
}
\description{
Get MERMAID endpoint
}
