% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_get_all_project_endpoints.R
\name{mermaid_get_all_project_endpoints}
\alias{mermaid_get_all_project_endpoints}
\title{Get all endpoints from specified MERMAID project(s)}
\usage{
mermaid_get_all_project_endpoints(
  project = mermaid_get_default_project(),
  limit = NULL,
  url = base_url,
  token = mermaid_token()
)
}
\arguments{
\item{project}{A way to identify the project(s). Can be project IDs (passed as a character vector directly) or projects resulting from \code{\link{mermaid_get_endpoint}} or \code{\link{mermaid_search_projects}}. Defaults to the projects listed via \code{get_default_project}, if available.}

\item{limit}{Number of records to get. Use NULL (the default) to get all records.}

\item{url}{API URL. Defaults to https://api.datamermaid.org}

\item{token}{API token. Not required for unauthenticated endpoints. Get via \code{\link{mermaid_auth}}}
}
\value{
A list of tibbles, one for each of the MERMAID project endpoints.
}
\description{
Get all endpoints from specified MERMAID project(s)
}
\examples{
\dontrun{
test_project <- mermaid_search_projects("Sharla test", include_test_projects = TRUE)
mermaid_get_all_project_endpoints(test_project, limit = 1)
}
}
