% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_get_project_endpoint.R
\name{mermaid_get_project_endpoint}
\alias{mermaid_get_project_endpoint}
\title{Get other MERMAID API project endpoints}
\usage{
mermaid_get_project_endpoint(
  project = mermaid_get_default_project(),
  endpoint = c("beltfishtransectmethods", "beltfishes", "benthiclittransectmethods",
    "benthicpittransectmethods", "benthicpits", "benthictransects", "collectrecords",
    "fishbelttransects", "habitatcomplexities", "obsbenthiclits", "obsbenthicpits",
    "obshabitatcomplexities", "obstransectbeltfishs", "observers", "project_profiles",
    "sampleevents"),
  limit = NULL,
  url = base_url,
  token = mermaid_token()
)
}
\arguments{
\item{project}{A way to identify the project(s). Can be project IDs (passed as a character vector directly) or projects resulting from \code{\link{mermaid_get_endpoint}} or \code{\link{mermaid_search_projects}}. Defaults to the projects listed via \code{get_default_project}, if available.}

\item{endpoint}{Endpoint}

\item{limit}{Number of records to get. Use NULL (the default) to get all records.}

\item{url}{API URL. Defaults to https://api.datamermaid.org}

\item{token}{API token. Not required for unauthenticated endpoints. Get via \code{\link{mermaid_auth}}}
}
\description{
Get other MERMAID API project endpoints
}
\examples{
projects <- mermaid_get_my_projects()
projects \%>\%
  mermaid_get_project_endpoint("observers")
}
