% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_get_project_gfcr_report.R
\name{mermaid_get_project_gfcr_report}
\alias{mermaid_get_project_gfcr_report}
\title{Get GFCR report for project}
\usage{
mermaid_get_project_gfcr_report(project, token = mermaid_token())
}
\arguments{
\item{project}{A way to identify the project(s). Can be project IDs (passed as a character vector directly) or projects resulting from \code{\link{mermaid_get_my_projects}} or \code{\link{mermaid_search_my_projects}}. Defaults to the projects listed via \code{mermaid_get_default_project}, if available.}

\item{token}{API token. Authenticate manually via \code{\link{mermaid_auth}}, or automatically when running any project- or user-specific functions (like this one).}
}
\description{
Get GFCR report for project
}
\examples{
\dontrun{
projects <- mermaid_get_my_projects()
projects \%>\%
  head(1) \%>\%
  mermaid_get_project_gfcr_report()
}
}
