% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_get_reference.R
\name{mermaid_get_reference}
\alias{mermaid_get_reference}
\title{Get MERMAID reference}
\usage{
mermaid_get_reference(
  reference = c("fishfamilies", "fishgenera", "fishspecies", "benthicattributes",
    "invertattributes", "invertspecies"),
  limit = NULL
)
}
\arguments{
\item{reference}{MERMAID reference. One of "fishfamilies", "fishgenera", "fishspecies", "benthicattributes", "invertattributes", "invertspecies".}

\item{limit}{Number of records to get. Use NULL (the default) to get all records.}
}
\description{
Find the names and information of the fish and benthic attributes you can choose in MERMAID.
}
\examples{
\dontrun{
mermaid_get_reference("benthicattributes")
mermaid_get_reference(c("fishfamilies", "fishgenera"))
}
}
