% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_get_sites.R
\name{mermaid_get_sites}
\alias{mermaid_get_sites}
\title{Get MERMAID sites}
\usage{
mermaid_get_sites(limit = NULL)
}
\arguments{
\item{limit}{Number of records to get. Use NULL (the default) to get all records.}
}
\description{
Get a data frame of all MERMAID sites (not project specific).
}
\examples{
\dontrun{
mermaid_get_sites(limit = 10)
}
}
