% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_get_summary_sites.R
\name{mermaid_get_summary_sites}
\alias{mermaid_get_summary_sites}
\title{Get aggregated metrics for all MERMAID surveys, by site, for all dates}
\usage{
mermaid_get_summary_sites(limit = NULL)
}
\arguments{
\item{limit}{Number of records to get. Use NULL (the default) to get all records.}
}
\value{

}
\description{
Get aggregated metrics from all surveys associated with each site, for all dates. Includes data from all methods - Fish Belt, Benthic LIT, Benthic PIT, Bleaching, and Habitat Complexity - if the data sharing policy for that method is public summary or public (and just includes the sample unit count otherwise). Does not require authorization.
}
\examples{
mermaid_get_summary_sites()
}
