% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_import_bulk_submit.R
\name{mermaid_import_bulk_submit}
\alias{mermaid_import_bulk_submit}
\title{Bulk submit valid records in Collecting}
\usage{
mermaid_import_bulk_submit(project, token = mermaid_token())
}
\arguments{
\item{project}{A way to identify the project(s). Can be project IDs (passed as a character vector directly) or projects resulting from \code{\link{mermaid_get_my_projects}} or \code{\link{mermaid_search_my_projects}}. Defaults to the projects listed via \code{mermaid_get_default_project}, if available.}

\item{token}{API token. Authenticate manually via \code{\link{mermaid_auth}}, or automatically when running any project- or user-specific functions (like this one).}
}
\description{
Bulk submit validated records in Collecting for a given project. Only submits records that have successfully validated without ANY errors or warnings. To be used after \code{\link{mermaid_import_project_data}} and \code{\link{mermaid_import_bulk_validate}}.
}
\examples{
\dontrun{
p <- mermaid_get_my_projects() \%>\%
  head(1)

p \%>\%
  mermaid_import_bulk_validate()

# 43 records being validated...
# ✖ 27 records produced errors in validation
# • 13 records produced warnings in validation
# ✔ 3 records successfully validated without warnings or errors

p \%>\%
  mermaid_import_bulk_submit()
# 3 records being submitted...
# ✔ 3 records successfully submitted
}
}
