% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_import_check_options.R
\name{mermaid_import_check_options}
\alias{mermaid_import_check_options}
\title{Check field options for MERMAID import}
\usage{
mermaid_import_check_options(data, options, field)
}
\arguments{
\item{data}{Data to be imported into MERMAID.}

\item{options}{Field options for the same method as \code{data}, from \code{\link{mermaid_import_get_options}}.}

\item{field}{Field to check options for (from \code{options}).}
}
\description{
Check that your data matches allowed field options for importing data for a given method into MERMAID.
}
\examples{
library(dplyr)

projects <- mermaid_get_my_projects()

options <- projects \%>\%
  head(1) \%>\%
  mermaid_import_get_options("fishbelt")

data <- tibble(Visibility = c("<1m - bad", "10+m - exellent"))

data \%>\%
  mermaid_import_check_options(options, "Visibility")

# • Some errors in values of `Visibility` - please check table below
# # A tibble: 2 × 3
# data_value        closest_choice   match
# <chr>             <chr>            <lgl>
# 1 10+m - exellent 10+m - excellent FALSE
# 2 <1m - bad       <1m - bad        TRUE
}
