% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_import_field_options.R
\name{mermaid_import_field_options}
\alias{mermaid_import_field_options}
\title{Get or check field options for MERMAID import}
\usage{
mermaid_import_field_options(
  field = c("width", "fishsizebin", "reefslope", "visibility", "current",
    "relativedepth", "tide", "fishname", "benthicattribute", "growthform",
    "habitatcomplexityscore"),
  matches
)
}
\arguments{
\item{field}{Field to check options for. One of: width, fishsizebin, reefslope, visibility, current, relativedepth, tide, fishname, benthicattribute, growthform, habitatcomplexityscore.}

\item{matches}{Optional options from your data, to match against valid options in Collect}
}
\description{
Check the options available for a given field in MERMAID, either to see what the options are or to check that your data matches the available fields.
}
\examples{
{
mermaid_import_field_options("growthform")
}
}
