% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_import_get_options.R
\name{mermaid_import_get_options}
\alias{mermaid_import_get_options}
\title{Get field options for MERMAID import}
\usage{
mermaid_import_get_options(
  project,
  method = c("fishbelt", "benthiclit", "benthicpit", "benthicpqt", "bleaching",
    "habitatcomplexity"),
  save,
  token = mermaid_token()
)
}
\arguments{
\item{project}{A way to identify the project(s). Can be project IDs (passed as a character vector directly) or projects resulting from \code{\link{mermaid_get_my_projects}} or \code{\link{mermaid_search_my_projects}}. Defaults to the projects listed via \code{mermaid_get_default_project}, if available.}

\item{method}{Method to get import fields and options for. One of "fishbelt", "benthiclit", "benthicpit", "benthicpqt", "bleaching", or "habitatcomplexity".}

\item{save}{Excel file to save field options to - .xlsx or xls file. Optional.}

\item{token}{API token. Authenticate manually via \code{\link{mermaid_auth}}, or automatically when running any project- or user-specific functions (like this one).}
}
\description{
Check the options available for importing a given method into MERMAID to see what the options are for each field. Returns a list of each field for a given method, whether it's required (in \code{required}), any available description or help with the field (in \code{help_text}), and a list of available choices, if relevant (in \code{choices}). Optionally, the fields and options can be saved into an Excel file using the \code{save} parameter. See also \code{\link{mermaid_import_check_options}} for checking that your data matches the allowed values.
}
\examples{
\dontrun{
options <- mermaid_get_my_projects() \%>\%
  head(1) \%>\%
  mermaid_import_get_options("fishbelt")

names(options)
# [1] "Site *"                      "Management *"
# [3] "Sample date: Year *"         "Sample date: Month *"
# [5] "Sample date: Day *"          "Sample time"
# [7] "Depth *"                     "Transect number *"
# [9] "Transect label"              "Transect length surveyed *"
# [11] "Width *"                    "Fish size bin *"
# [13] "Reef slope"                 "Visibility"
# [15] "Current"                    "Relative depth"
# [17] "Tide"                       "Sample unit notes"
# [19] "Observer emails *"          "Fish name *"
# [21] "Size *"                     "Count *"

options[["Reef slope"]][["required"]]
# [1] FALSE

options[["Reef slope"]][["choices"]]
# A tibble: 4 × 1
# value
# <chr>
# 1 crest
# 2 flat
# 3 slope
# 4 wall
}
}
