% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_import_get_template.R
\name{mermaid_import_get_template}
\alias{mermaid_import_get_template}
\title{Get template for MERMAID import}
\usage{
mermaid_import_get_template(
  method = c("fishbelt", "benthiclit", "benthicpit", "benthicpqt", "bleachingqc",
    "habitatcomplexity", "all"),
  save
)
}
\arguments{
\item{method}{Method to get import template for. One of "fishbelt", "benthiclit", "benthicpit", "benthicpqt", "bleaching", "habitatcomplexity", or "all" (to get templates for all methods).}

\item{save}{Excel file to save template to. Optional.}
}
\description{
Get a template of the fields for importing into MERMAID. Used along with \code{\link{mermaid_import_get_options}}, which contains details on which fields are required and their possible options, if relevant. Optionally, the template can be saved into an Excel file using the \code{save} parameter.
}
