% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mermaid_projects.R
\name{mermaid_projects}
\alias{mermaid_projects}
\title{Get MERMAID Projects}
\usage{
mermaid_projects(limit = 50)
}
\arguments{
\item{limit}{Number of projects to get. Defaults to 50.}
}
\description{
Get MERMAID projects. Returns the projects
}
\examples{
\dontrun{
mermaid_projects()
}
}
