% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_search_projects.R
\name{mermaid_search_projects}
\alias{mermaid_search_projects}
\title{Search projects by name or ID}
\usage{
mermaid_search_projects(name = NULL, id = NULL)
}
\arguments{
\item{name}{Project name}

\item{id}{Project ID}
}
\description{
Returns metadata on project, including ID, name, countries, number of sites, tags, notes, status, data sharing policies, and when the project was created and last updated.
}
