% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_get.R
\name{project_belt_fish_transect_methods}
\alias{project_belt_fish_transect_methods}
\alias{project_beltfishes}
\alias{project_benthic_lit_transect_methods}
\alias{project_benthic_pit_transect_methods}
\alias{project_benthic_pits}
\alias{project_benthic_transects}
\alias{project_collect_records}
\alias{project_fishbelttransects}
\alias{project_habitat_complexities}
\alias{project_obs_benthic_lits}
\alias{project_obs_benthic_pits}
\alias{project_obs_habitat_complexities}
\alias{project_obs_transect_beltfishs}
\alias{project_managements}
\alias{project_observers}
\alias{project_profiles}
\alias{project_project_profiles}
\alias{project_sample_events}
\alias{project_sites}
\title{Get endpoints for a given project.}
\usage{
project_belt_fish_transect_methods(project_id, limit = 50,
  results_only = TRUE)

project_beltfishes(project_id, limit = 50, results_only = TRUE)

project_benthic_lit_transect_methods(project_id, limit = 50,
  results_only = TRUE)

project_benthic_pit_transect_methods(project_id, limit = 50,
  results_only = TRUE)

project_benthic_pits(project_id, limit = 50, results_only = TRUE)

project_benthic_transects(project_id, limit = 50, results_only = TRUE)

project_collect_records(project_id, limit = 50, results_only = TRUE)

project_fishbelttransects(project_id, limit = 50, results_only = TRUE)

project_habitat_complexities(project_id, limit = 50,
  results_only = TRUE)

project_obs_benthic_lits(project_id, limit = 50, results_only = TRUE)

project_obs_benthic_pits(project_id, limit = 50, results_only = TRUE)

project_obs_habitat_complexities(project_id, limit = 50,
  results_only = TRUE)

project_obs_transect_beltfishs(project_id, limit = 50,
  results_only = TRUE)

project_managements(project_id, limit = 50, results_only = TRUE)

project_observers(project_id, limit = 50, results_only = TRUE)

project_profiles(project_id, limit = 50, results_only = TRUE)

project_project_profiles(project_id, limit = 50, results_only = TRUE)

project_sample_events(project_id, limit = 50, results_only = TRUE)

project_sites(project_id, limit = 50, results_only = TRUE)
}
\arguments{
\item{project_id}{Project ID}

\item{limit}{Number of records to get. Defaults to 50.}

\item{results_only}{Whether to return the results only (the default) or to return the entire API call (contents, results, path, and response).}
}
\description{
These wrapper functions make it easy to get data for a given project.
There are 19 functions:
\itemize{
  \item \code{project_belt_fish_transect_methods()}
  \item \code{project_beltfishes()}
  \item \code{project_benthic_lit_transect_methods()}
  \item \code{project_benthic_pit_transect_methods()}
  \item \code{project_benthic_pits()}
  \item \code{project_benthic_transects()}
  \item \code{project_collect_records()}
  \item \code{project_fishbelttransects()}
  \item \code{project_habitat_complexities()}
  \item \code{project_obs_benthic_lits()}
  \item \code{project_obs_benthic_pits()}
  \item \code{project_obs_habitat_complexities()}
  \item \code{project_obs_transect_beltfishs()}
  \item \code{project_managements()}
  \item \code{project_observers()}
  \item \code{project_profiles()}
  \item \code{project_project_profiles()}
  \item \code{project_sample_events()}
  \item \code{project_sites()}

}
}
\examples{
print("todo")
}
