% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_projects.R
\name{search_projects}
\alias{search_projects}
\title{Search projects by name or ID}
\usage{
search_projects(name = NULL, id = NULL, fixed = FALSE)
}
\arguments{
\item{name}{Project name}

\item{id}{Project ID}
}
\description{
Returns metadata on project, including ID, name, countries, number of sites, tags, notes, status, data sharing policies, and when the project was created and last updated.
}
\examples{
\dontrun {
search_projects(name = "mermaidr testing")
}
}
