% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_countries.R
\name{mermaid_countries}
\alias{mermaid_countries}
\title{MERMAID countries}
\usage{
mermaid_countries(url = base_url)
}
\arguments{
\item{url}{API URL. Defaults to https://api.datamermaid.org}
}
\description{
Returns the names of countries in MERMAID to help with searching.
}
\examples{
\dontrun{
mermaid_countries()
}
}
