% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_get.R
\name{mermaid_get}
\alias{mermaid_get}
\title{Get MERMAID endpoint}
\usage{
mermaid_get(endpoint, limit = 50)
}
\arguments{
\item{endpoint}{Endpoint}

\item{limit}{Number of records to get. Defaults to 50.}
}
\description{
Get MERMAID endpoint
}
