% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_get.R
\name{mermaid_GET}
\alias{mermaid_GET}
\title{Get MERMAID endpoint}
\usage{
mermaid_GET(endpoint, limit = NULL, token = NULL, ...)
}
\arguments{
\item{endpoint}{Endpoint}

\item{limit}{Number of records to get. Use NULL (the default) to get all records.}

\item{token}{API token. Not required for unauthenticated endpoints. Get via \code{\link{mermaid_auth}}}

\item{...}{Additional parameters used as needed}
}
\description{
Get MERMAID endpoint. You should not need to call this function directly - please use the `mermaid_get_*` functions instead.
}
