% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_get_classification_labelmappings.R
\name{mermaid_get_classification_labelmappings}
\alias{mermaid_get_classification_labelmappings}
\title{Classification label mappings}
\usage{
mermaid_get_classification_labelmappings(limit = NULL)
}
\arguments{
\item{limit}{Number of records to get. Use NULL (the default) to get all records.}
}
\description{
Get a lookup table that maps MERMAID Benthic Attributes and Growth Forms to labels from other sources (e.g. CoralNet).
}
\examples{
mermaid_get_classification_labelmappings()
}
