% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_get_endpoint.R
\name{mermaid_get_endpoint}
\alias{mermaid_get_endpoint}
\title{Get other MERMAID API endpoints}
\usage{
mermaid_get_endpoint(
  endpoint = c("choices", "projecttags", "fishsizes"),
  limit = NULL,
  url = base_url
)
}
\arguments{
\item{endpoint}{MERMAID API endpoint. One of "choices", "projecttags", "fishsizes".}

\item{limit}{Number of records to get. Use NULL (the default) to get all records.}

\item{url}{API URL. Defaults to https://api.datamermaid.org}
}
\description{
Get other MERMAID API endpoints not covered by other \code{mermaid_get_*} functions.
}
\examples{
\dontrun{
mermaid_get_endpoint("choices")
}
}
