% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_get_endpoint.R
\name{mermaid_get_endpoint}
\alias{mermaid_get_endpoint}
\title{Get MERMAID Endpoint}
\usage{
mermaid_get_endpoint(
  endpoint = c("benthicattributes", "choices", "fishfamilies", "fishgenera",
    "fishspecies", "fishsizes", "managements", "projects", "projecttags", "sites"),
  limit = NULL,
  url = base_url,
  ...
)
}
\arguments{
\item{endpoint}{Endpoint}

\item{limit}{Number of records to get. Use NULL (the default) to get all records.}

\item{url}{API URL. Defaults to https://api.datamermaid.org}

\item{...}{Additional parameters used as needed}
}
\description{
Get MERMAID Endpoint
}
\examples{
mermaid_get_endpoint("sites", limit = 1)
}
