% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_get_managements.R
\name{mermaid_get_managements}
\alias{mermaid_get_managements}
\title{Get MERMAID managements}
\usage{
mermaid_get_managements(limit = NULL, url = base_url)
}
\arguments{
\item{limit}{Number of records to get. Use NULL (the default) to get all records.}

\item{url}{API URL. Defaults to https://api.datamermaid.org}
}
\description{
Get a data frame of all MERMAID managements (not project specific).
}
\examples{
\dontrun{
mermaid_get_managements(limit = 10)
}
}
