% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_get_me.R
\name{mermaid_get_me}
\alias{mermaid_get_me}
\title{Get your MERMAID Profile}
\usage{
mermaid_get_me(token = mermaid_token(), field_report = TRUE)
}
\arguments{
\item{token}{API token. Not required for unauthenticated endpoints. Get manually via \code{\link{mermaid_auth}} or automatically when running a function that requires a token.}

\item{field_report}{Whether the output should be a "field report", i.e. a cleaned up version of the output with IDs mapped to names and only relevant fields shown. Defaults to TRUE, and setting to FALSE is not recommended outside of advanced usage.}
}
\value{
A tibble of MERMAID profile information, including first and last name, full name, and email address.
}
\description{
Get your MERMAID profile. Returns ID, first/last name, and email address. Requires authorization.
}
\examples{
\dontrun{
mermaid_get_me()
}
}
