% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_get_my_projects.R
\name{mermaid_get_my_projects}
\alias{mermaid_get_my_projects}
\title{Get a list of your MERMAID Projects}
\usage{
mermaid_get_my_projects(
  include_test_projects = FALSE,
  limit = NULL,
  url = base_url,
  token = mermaid_token()
)
}
\arguments{
\item{include_test_projects}{Whether to include test projects. Defaults to FALSE.}

\item{limit}{Number of records to get. Use NULL (the default) to get all records.}

\item{url}{API URL. Defaults to https://api.datamermaid.org}

\item{token}{API token. Not required for unauthenticated endpoints. Get via \code{\link{mermaid_auth}}}
}
\value{
A tibble of MERMAID projects that you have access to, including project name, countries, number of sites, tags, notes, and data policies.
}
\description{
List MERMAID projects that you have access to. Requires authorization.
}
\examples{
\dontrun{
mermaid_get_my_projects(limit = 5)
}
}
