% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_get_my_projects.R
\name{mermaid_get_my_projects}
\alias{mermaid_get_my_projects}
\title{Get your MERMAID Projects}
\usage{
mermaid_get_my_projects(
  include_test_projects = FALSE,
  limit = NULL,
  token = mermaid_token()
)
}
\arguments{
\item{include_test_projects}{Whether to include test projects. Defaults to FALSE.}

\item{limit}{Number of records to get. Use NULL (the default) to get all records.}

\item{token}{API token. Not required for unauthenticated endpoints (non project- or user-specific). Get manually via \code{\link{mermaid_auth}} or automatically when running a project- or user-specific function.}
}
\value{
A tibble of MERMAID projects that you have access to, including project name, countries, number of sites, tags, notes, and data policies.
}
\description{
Get MERMAID projects you have access to. Returns metadata on projects, including ID, name, countries, number of sites, tags, notes, status, data sharing policies, and when the project was created and last updated. Requires authorization.
}
\examples{
\dontrun{
mermaid_get_my_projects()
}
}
