% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_get_project_data.R
\name{mermaid_get_project_data}
\alias{mermaid_get_project_data}
\title{Get MERMAID project data}
\usage{
mermaid_get_project_data(
  project,
  method = c("beltfishes", "benthicpits"),
  data = c("observations", "sampleunits", "sampleevents"),
  limit = NULL,
  url = base_url,
  token = mermaid_token(),
  ...
)
}
\arguments{
\item{project}{A way to identify the project(s). Can be project IDs (passed as a character vector directly) or projects resulting from \code{\link{mermaid_get_endpoint}} or \code{\link{mermaid_search_projects}}. Defaults to the projects listed via \code{get_default_project}, if available.}

\item{method}{Method to get data for. One of "beltfishes", "benthicpits", or "all" (to get data for both methods).}

\item{data}{Data to return. One of "observations", "sampleunits", "sampleevents", or all (to get all three kinds of data). See details for more.}

\item{limit}{Number of records to get. Use NULL (the default) to get all records.}

\item{url}{API URL. Defaults to https://api.datamermaid.org}

\item{token}{API token. Not required for unauthenticated endpoints. Get via \code{\link{mermaid_auth}}}
}
\description{
Get MERMAID project data
}
\examples{
projects <- mermaid_get_my_projects()
projects \%>\%
  mermaid_get_project_data(method = "beltfishes", data = "sampleevents", limit = 10)

projects \%>\%
  mermaid_get_project_data(method = c("benthicpits", "beltfishes"), data = "sampleevents", limit = 10)
}
