% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_get_project_endpoint.R
\name{mermaid_get_project_endpoint}
\alias{mermaid_get_project_endpoint}
\title{Get other MERMAID API project endpoints}
\usage{
mermaid_get_project_endpoint(
  project = mermaid_get_default_project(),
  endpoint = c("beltfishtransectmethods", "benthiclittransectmethods",
    "benthicpittransectmethods", "benthictransects", "collectrecords",
    "fishbelttransects", "observers", "project_profiles", "sampleevents"),
  limit = NULL,
  token = mermaid_token(),
  as_is = FALSE
)
}
\arguments{
\item{project}{A way to identify the project(s). Can be project IDs (passed as a character vector directly) or projects resulting from \code{\link{mermaid_get_my_projects}} or \code{\link{mermaid_search_my_projects}}. Defaults to the projects listed via \code{mermaid_get_default_project}, if available.}

\item{endpoint}{Endpoint}

\item{limit}{Number of records to get. Use NULL (the default) to get all records.}

\item{token}{API token. Authenticate manually via \code{\link{mermaid_auth}}, or automatically when running any project- or user-specific functions (like this one).}

\item{as_is}{Whether the output should be a "as is", i.e. a cleaned up version of the output with IDs mapped to names and only relevant fields shown. Defaults to FALSE, and setting to TRUE is not recommended outside of advanced usage.}
}
\description{
Get data from MERMAID API project endpoints not covered by other \code{mermaid_get_project_*} functions. Requires authorization.
}
\examples{
\dontrun{
projects <- mermaid_get_my_projects()
projects \%>\%
  mermaid_get_project_endpoint("observers")
}
}
