% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_get_project_sites.R
\name{mermaid_get_project_sites}
\alias{mermaid_get_project_sites}
\title{Get MERMAID project sites}
\usage{
mermaid_get_project_sites(
  project = mermaid_get_default_project(),
  limit = NULL,
  url = base_url,
  token = mermaid_token()
)
}
\arguments{
\item{project}{A way to identify the project(s). Can be project IDs (passed as a character vector directly) or projects resulting from \code{\link{mermaid_get_my_projects}} or \code{\link{mermaid_search_my_projects}}. Defaults to the projects listed via \code{mermaid_get_default_project}, if available.}

\item{limit}{Number of records to get. Use NULL (the default) to get all records.}

\item{url}{API URL. Defaults to https://api.datamermaid.org}

\item{token}{API token. Not required for unauthenticated endpoints. Get via \code{\link{mermaid_auth}}}
}
\description{
Get sites for specified MERMAID project(s).
}
\examples{
\dontrun{
projects <- mermaid_get_my_projects()
projects \%>\%
  mermaid_get_project_sites()
}
}
