% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_get_project_sites.R
\name{mermaid_get_project_sites}
\alias{mermaid_get_project_sites}
\title{Get MERMAID project sites}
\usage{
mermaid_get_project_sites(
  project = mermaid_get_default_project(),
  limit = NULL,
  token = mermaid_token(),
  covariates = FALSE
)
}
\arguments{
\item{project}{A way to identify the project(s). Can be project IDs (passed as a character vector directly) or projects resulting from \code{\link{mermaid_get_my_projects}} or \code{\link{mermaid_search_my_projects}}. Defaults to the projects listed via \code{mermaid_get_default_project}, if available.}

\item{limit}{Number of records to get. Use NULL (the default) to get all records.}

\item{token}{API token. Authenticate manually via \code{\link{mermaid_auth}}, or automatically when running any project- or user-specific functions (like this one).}
}
\description{
Get sites for specified MERMAID project(s). Returns metadata on sites, including site ID and name, notes, latitude and longitude, country, reef type and zone, exposure, and when the site was created and last updated. Optionally, get covariates (Allen Coral Atlas, Andrello, Beyer) for sites. Requires authorization.
}
\examples{
\dontrun{
projects <- mermaid_get_my_projects()
projects \%>\%
  mermaid_get_project_sites()
}
}
