% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_get_projects.R
\name{mermaid_get_projects}
\alias{mermaid_get_projects}
\title{Get a list of all MERMAID Projects}
\usage{
mermaid_get_projects(include_test_projects = FALSE, limit = NULL)
}
\arguments{
\item{include_test_projects}{Whether to include test projects. Defaults to FALSE.}

\item{limit}{Number of records to get. Use NULL (the default) to get all records.}
}
\value{
A tibble of MERMAID projects including project name, countries, number of sites, tags, notes, and data policies.
}
\description{
Get a list of all MERMAID Projects
}
\examples{
\dontrun{
mermaid_get_projects(limit = 5)
}
}
