% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_get_projects.R
\name{mermaid_get_projects}
\alias{mermaid_get_projects}
\title{Get all MERMAID Projects}
\usage{
mermaid_get_projects(include_test_projects = FALSE, limit = NULL)
}
\arguments{
\item{include_test_projects}{Whether to include test projects. Defaults to FALSE.}

\item{limit}{Number of records to get. Use NULL (the default) to get all records.}
}
\description{
Get all MERMAID projects. Returns metadata on projects, including ID, name, countries, number of sites, tags, notes, status, data sharing policies, and when the project was created and last updated.
}
\examples{
\dontrun{
mermaid_get_projects(limit = 5)
}
}
