% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_get_reference.R
\name{mermaid_get_reference}
\alias{mermaid_get_reference}
\title{Get MERMAID reference}
\usage{
mermaid_get_reference(
  reference = c("fishfamilies", "fishgenera", "fishspecies", "benthicattributes"),
  limit = NULL,
  field_report = TRUE
)
}
\arguments{
\item{reference}{MERMAID reference. One of "fishfamilies", "fishgenera", "fishspecies", "benthicattributes".}

\item{limit}{Number of records to get. Use NULL (the default) to get all records.}

\item{field_report}{Whether the output should be a "field report", i.e. a cleaned up version of the output with IDs mapped to names and only relevant fields shown. Defaults to TRUE, and setting to FALSE is not recommended outside of advanced usage.}
}
\description{
Find the names and information of the fish and benthic attributes you can choose in MERMAID.
}
\examples{
\dontrun{
mermaid_get_reference("benthicattributes")
mermaid_get_reference(c("fishfamilies", "fishgenera"))
}
}
