% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_get_sites.R
\name{mermaid_get_sites}
\alias{mermaid_get_sites}
\title{Get MERMAID sites}
\usage{
mermaid_get_sites(limit = NULL)
}
\arguments{
\item{limit}{Number of records to get. Use NULL (the default) to get all records.}
}
\description{
Get all MERMAID sites (not project specific). Returns metadata on sites, including site ID and name, notes, project, latitude and longitude, country, reef type and zone, exposure, and when the site was created and last updated.
}
\examples{
\dontrun{
mermaid_get_sites(limit = 10)
}
}
