% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_get_summary_sampleevents.R
\name{mermaid_get_summary_sampleevents}
\alias{mermaid_get_summary_sampleevents}
\title{Get aggregated metrics for all MERMAID surveys, by site, by date}
\usage{
mermaid_get_summary_sampleevents(limit = NULL)
}
\arguments{
\item{limit}{Number of records to get. Use NULL (the default) to get all records.}
}
\description{
Get aggregated metrics from all surveys associated with each site, by date. Includes data from all methods - Fish Belt, Benthic LIT, Benthic PIT, Bleaching, and Habitat Complexity - if the data sharing policy for that method is public summary or public (and just includes the sample unit count otherwise). Does not require authorization.
}
\examples{
\dontrun{
mermaid_get_summary_sampleevents()
}
}
