% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_import_bulk_edit.R
\name{mermaid_import_bulk_edit}
\alias{mermaid_import_bulk_edit}
\title{Bulk edit submitted records, moving them back into Collecting}
\usage{
mermaid_import_bulk_edit(project, method = NULL, token = mermaid_token())
}
\arguments{
\item{project}{A way to identify the project(s). Can be project IDs (passed as a character vector directly) or projects resulting from \code{\link{mermaid_get_my_projects}} or \code{\link{mermaid_search_my_projects}}. Defaults to the projects listed via \code{mermaid_get_default_project}, if available.}

\item{method}{Method to return submitted data back into "editing" state for. One of "fishbelt", "benthiclit", "benthicpit", "benthicpqt", bleaching", or "habitatcomplexity".}

\item{token}{API token. Authenticate manually via \code{\link{mermaid_auth}}, or automatically when running any project- or user-specific functions (like this one).}
}
\description{
Returns submitted records back to Collecting, where they can be edited, for a given project and method. This should only be used when errors in data are discovered and ALL records need to be moved back to Collecting. To be used after \code{\link{mermaid_import_project_data}}, \code{\link{mermaid_import_bulk_validate}}, and \code{\link{mermaid_import_bulk_submit}}.
}
\examples{
\dontrun{
p <- mermaid_get_my_projects() \%>\%
  head(1)

p \%>\%
  mermaid_import_bulk_validate()

# 43 records being validated...
# ✖ 27 records produced errors in validation
# • 13 records produced warnings in validation
# ✔ 3 records successfully validated without warnings or errors

p \%>\%
  mermaid_import_bulk_submit()
# 3 records being submitted...
# ✔ 3 records successfully submitted

mermaid_import_bulk_edit(p, "fishbelt")
✔ 3 record successfully edited and moved back to Collecting.
}
}
