% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_import_project_data.R
\name{mermaid_import_project_data}
\alias{mermaid_import_project_data}
\title{Import data into MERMAID Collect}
\usage{
mermaid_import_project_data(
  data,
  project_id,
  method = c("fishbelt", "benthicpit", "benthiclit", "habitatcomplexity", "bleaching"),
  dryrun = TRUE,
  token = mermaid_token()
)
}
\arguments{
\item{data}{Data to import. Either a data frame or a file path.}

\item{project_id}{ID of project to import data into.}

\item{method}{Method to import data for. One of "fishbelt", "benthiclit", "benthicpit", "bleaching", "habitatcomplexity"}

\item{dryrun}{Whether the upload is a dry run, defaults to \code{TRUE}. If \code{TRUE}, records are validated and valid Collect records are created but NOT saved to the database. If \code{FALSE}, records are validated and valid Collect reocrds are saved to the database.}

\item{token}{API token.}
}
\description{
Import data into MERMAID Collect. If successful, the function returns a message saying that it was successful with a link to review the records in Collect. If unsuccessful, the function returns a data frame of validation errors.
}
