% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_import_project_managements.R
\name{mermaid_import_project_managements}
\alias{mermaid_import_project_managements}
\title{Import project managements into MERMAID Collect}
\usage{
mermaid_import_project_managements(project, data, token = mermaid_token())
}
\arguments{
\item{project}{A way to identify the project(s). Can be project IDs (passed as a character vector directly) or projects resulting from \code{\link{mermaid_get_my_projects}} or \code{\link{mermaid_search_my_projects}}. Defaults to the projects listed via \code{mermaid_get_default_project}, if available.}

\item{data}{Data to import. A data frame containing the fields: \code{name}, \code{latitude}, \code{longitude}, \code{country}, \code{reef_type}, \code{reef_zone}, \code{exposure}, and optionally \code{notes}.}

\item{token}{API token. Authenticate manually via \code{\link{mermaid_auth}}, or automatically when running any project- or user-specific functions (like this one).}
}
\description{
Import project managements into MERMAID Collect
}
\examples{
\dontrun{
project <- mermaid_search_my_projects("Sharla test", include_test_projects = TRUE)

data <- dplyr::tibble(
  name = "Test site",
  latitude = 43.651070,
  longitude = -79.347015,
  notes = NA,
  country = "Canada",
  reef_type = "fringing",
  reef_zone = "crest",
  exposure = "semi-exposed"
)

project \%>\%
  mermaid_import_project_managements(data)
}
}
