% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_list_projects.R
\name{mermaid_list_projects}
\alias{mermaid_list_projects}
\title{List MERMAID Projects}
\usage{
mermaid_list_projects(limit = 50, url = base_url)
}
\arguments{
\item{limit}{Number of records to get. Defaults to 50.}

\item{url}{API URL. Defaults to https://api.datamermaid.org}
}
\value{
A tibble of MERMAID projects including project name, countries, number of sites, tags, notes, and data policies.
}
\description{
List MERMAID Projects
}
\examples{
mermaid_list_projects(limit = 5)
}
