% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_list_projects.R
\name{mermaid_list_projects}
\alias{mermaid_list_projects}
\title{List MERMAID Projects}
\usage{
mermaid_list_projects(include_test_projects = FALSE, limit = NULL,
  url = base_url)
}
\arguments{
\item{include_test_projects}{Whether to include test projects. Defaults to FALSE.}

\item{limit}{Number of records to get. Use NULL (the default) to get all records.}

\item{url}{API URL. Defaults to https://api.datamermaid.org}
}
\value{
A tibble of MERMAID projects including project name, countries, number of sites, tags, notes, and data policies.
}
\description{
List MERMAID Projects
}
\examples{
mermaid_list_projects(limit = 5)
}
