% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_search_my_projects.R
\name{mermaid_search_my_projects}
\alias{mermaid_search_my_projects}
\title{Search your MERMAID projects}
\usage{
mermaid_search_my_projects(
  name = NULL,
  country = NULL,
  tag = NULL,
  include_test_projects = FALSE,
  limit = NULL,
  url = base_url,
  token = mermaid_token()
)
}
\arguments{
\item{name}{Project name}

\item{include_test_projects}{Whether to include test projects. Defaults to FALSE.}

\item{limit}{Number of records to get. Use NULL (the default) to get all records.}

\item{url}{API URL. Defaults to https://api.datamermaid.org}

\item{token}{API token. Not required for unauthenticated endpoints. Get via \code{\link{mermaid_auth}}}
}
\description{
Search within your MERMAID projects specfically. Returns metadata on project, including ID, name, countries, number of sites, tags, notes, status, data sharing policies, and when the project was created and last updated.
}
\examples{
\donttest{
mermaid_search_my_projects(tag = "WCS Indonesia")
mermaid_search_my_projects(countries = "Indonesia", tags = "Vibrant Oceans")

# The country (or tag) do not have to be exactly the same
# A project is returned if it *contains* the country/tag:
mermaid_search_my_projects(tags = "WCS", limit = 1)[["tags"]]
# [1] "WCS Indonesia; Vibrant Oceans"
}
}
