% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_search_projects.R
\name{mermaid_search_projects}
\alias{mermaid_search_projects}
\title{Search all MERMAID projects}
\usage{
mermaid_search_projects(
  name = NULL,
  countries = NULL,
  tags = NULL,
  include_test_projects = FALSE,
  limit = NULL,
  field_report = TRUE
)
}
\arguments{
\item{name}{Project name}

\item{countries}{Project country. Projects are returned if the \code{countries} field contains \code{countries}, not just if it is exactly the same. For a list of countries used in MERMAID, see \code{\link{mermaid_countries}}}

\item{tags}{Project tags. Projects are returned if the \code{tags} field contains \code{tags}, not just if it is exactly the same.}

\item{include_test_projects}{Whether to include test projects. Defaults to FALSE.}

\item{limit}{Number of records to get. Use NULL (the default) to get all records.}

\item{field_report}{Whether the output should be a "field report", i.e. a cleaned up version of the output with IDs mapped to names and only relevant fields shown. Defaults to TRUE, and setting to FALSE is not recommended outside of advanced usage.}
}
\description{
Search all MERMAID projects by project name, country, or tags. Returns metadata on project, including ID, name, countries, number of sites, tags, notes, status, data sharing policies, and when the project was created and last updated.
}
\examples{
\dontrun{
mermaid_search_projects(tags = "WCS Fiji")
mermaid_search_projects(countries = "Fiji", tags = "WWF-UK")

# The countries (or tags) do not have to be exactly the same
# A project is returned if it *contains* the countries/tags:
mermaid_search_projects(countries = "Tanzania", limit = 1)[["countries"]]

# To search within your projects only, use mermaid_search_my_projects():
mermaid_search_my_projects(countries = "Fiji")
}
}
