% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_project.R
\name{set_default_project}
\alias{set_default_project}
\alias{get_default_project}
\title{Get or set mermaidr MERMAID settings}
\usage{
set_default_project(project)

get_default_project()
}
\arguments{
\item{project}{A way to identify a MERMAID project. Can be a project ID (passed as a character vector directly) or a single project resulting from \code{\link{get_mermaid_endpoint}} or \code{\link{search_projects}}.}
}
\description{
Configure default MERMAID settings
}
\examples{
\dontrun{
test_project <- search_projects("mermaidr testing", exact_name = TRUE)
set_default_project(test_project)
get_default_project()
get_mermaid_project_endpoint("sites")
}
}
