% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_get.R
\name{project_get}
\alias{project_get}
\title{Get endpoint from a specified MERMAID project.}
\usage{
project_get(project_id, endpoint, limit = 50)
}
\arguments{
\item{project_id}{Project ID}

\item{endpoint}{API endpoint}

\item{limit}{Number of records to get. Defaults to 50.}
}
\description{
Get endpoint from a specified MERMAID project.
}
\examples{
\dontrun{

project_name <- "Beta testing"
projects <- mermaid_projects()
project_id <- dplyr::filter(projects, name == project_name)[["id"]]
project_get(project_id, "sampleevents")
}
}
