% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_mermaidr.R
\name{update_mermaidr}
\alias{update_mermaidr}
\title{Update `mermaidr`}
\usage{
update_mermaidr()
}
\description{
Update `mermaidr` if you do not have the latest version. Once the update is done, you must restart R.
}
